requirements.txt
absl-py==2.4.0
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.4.3
cupy-cuda12x==14.0.1
cycler==0.12.1
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.2.0
idna==3.11
imagededup==0.3.3.post2
ImageIO==2.37.3
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.33
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==25.0
pillow==12.1.1
pip==26.0.1
platformdirs==4.9.4
polars==1.39.3
polars-runtime-32==1.39.3
psutil==7.2.2
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyWavelets==1.9.0
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==80.10.2
six==1.17.0
sounddevice==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2026.3.3
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.3
triton==3.6.0
typing_extensions==4.15.0
ultralytics==8.4.24
ultralytics-thop==2.0.18
urllib3==2.6.3
wheel==0.46.3
wandb
